+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [36.4.0] 2024-07-29
+## [36.5.0] 2024-07-29
 ### Added
 - Support for asynchronous FHIRPath expressions.
+
+## [36.4.0] 2024-07-23
+### Added
+- Support for rendering-xhtml on item.text.
+- Added new template option allowHTML. allowHTMLInInstructions is deprecated.
 
 ## [36.3.3] 2024-07-22
 ### Added